requirements.txt
Django~=5.0.2
psycopg2-binary